+++ packages/dsw-database/CHANGELOG.md
@@ -7,6 +7,9 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Fixed
+
+- Stopped swallowing errors when retrieving mail config
 
 ## [4.33.0]