+++ packages/dsw-database/CHANGELOG.md
@@ -7,6 +7,9 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+
+## [4.34.0]
+
 ### Fixed
 
 - Stopped swallowing errors when retrieving mail config
@@ -15,7 +18,6 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Added
 
 - `PostgresConnection.discard` to give up a connection that may still be in use by an abandoned thread
-
 ## [4.33.0]
 
 Released for version consistency with other DSW tools.
@@ -519,3 +521,4 @@ Released for version consistency with other DSW tools.
 [4.31.0]: /../../tree/v4.31.0
 [4.32.0]: /../../tree/v4.32.0
 [4.33.0]: /../../tree/v4.33.0
+[4.34.0]: /../../tree/v4.34.0